requirements.txt
blinker==1.9.0
click==8.1.3
Flask==3.1.0
flask-peewee==3.0.6
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.18.1
python-dotenv==0.21.0
Werkzeug==3.1.3
wtf-peewee==3.0.6
WTForms==3.2.1
